Fix Custom Content frontstage.
