+++ e2e-tests/tests/Utils.ts
@@ -243,7 +243,7 @@ export async function openComponentExamples(
   page: Page,
   baseURL: string | undefined
 ) {
-  await page.goto(`${baseURL}/blank?frontstageId=widget-api`);
+  await page.goto(`${baseURL}/blank?menu=0`);
   await page.locator(".nz-toolbar-button-button").click();
   await page.getByRole("menuitem", { name: "Component Examples" }).click();
 }
